peyk/requirements.txt
fastapi==0.97.0
mysql-connector-python==8.0.33
pydantic==1.10.9
uvicorn==0.22.0
